Security and structure revised recording to copilot review
