+++ server/db/modules/import_degreeprogrammes.ts
@@ -1,83 +1,14 @@
 import fs from "fs";
 import { parseStringPromise } from "xml2js";
 import { Client } from "pg";
-import dotenv from "dotenv";
 import path from "path";
+import { getTextDe, batchInsert } from "./xml_import_helpers";
 
-dotenv.config();
-
-// Extraxt german text from possible multilingual fields
-function getTextDe(nameField: any): string | null {
-  if (!nameField) return null;
-
-  const namesArray = Array.isArray(nameField) ? nameField : [nameField];
-
-  for (const n of namesArray) {
-    if (n?.lang === "de" && typeof n._ === "string") {
-      return n._.trim();
-    }
-    if (n?.lang === "de" && typeof n === "string") {
-      return n.trim();
-    }
-  }
-  return null;
-}
-
-// Helper for batch-inserts
-const BATCH_SIZE = 500;
-
-async function batchInsert(
-  client: Client,
-  table: string,
-  columns: string[],
-  rows: any[][],
-  conflict?: string,
-) {
-  if (!rows.length) return;
-
-  for (let i = 0; i < rows.length; i += BATCH_SIZE) {
-    const batch = rows.slice(i, i + BATCH_SIZE);
-
-    const values: string[] = [];
-    const params: any[] = [];
-    let paramIndex = 1;
-
-    for (const row of batch) {
-      if (row.length !== columns.length) {
-        throw new Error(
-          `Row length ${row.length} does not match columns length ${columns.length}`,
-        );
-      }
-      const placeholders = row.map(() => `$${paramIndex++}`);
-      values.push(`(${placeholders.join(",")})`);
-      params.push(...row);
-    }
-
-    const query = `
-      INSERT INTO ${table} (${columns.join(",")})
-      VALUES ${values.join(",")}
-      ${conflict ? `ON CONFLICT ${conflict} DO NOTHING` : ""}
-    `;
-    await client.query(query, params);
-  }
-}
-
-async function main() {
-  const client = new Client({
-    user: process.env.DB_USER,
-    host: process.env.DB_HOST,
-    database: process.env.DB_NAME,
-    password: process.env.DB_PASSWORD,
-    port: Number(process.env.DB_PORT),
-  });
-
-  await client.connect();
-
+export async function importDegreeProgrammes(client: Client) {
   const xmlPath = path.join(__dirname, "../xml/degreeprogrammes.xml");
 
   if (!fs.existsSync(xmlPath)) {
-    console.error("XML file not found:", xmlPath);
-    process.exit(1);
+    throw new Error(`XML file not found: ${xmlPath}`);
   }
 
   const xmlData = fs.readFileSync(xmlPath, "utf-8");
@@ -102,8 +33,7 @@ async function main() {
     );
   }
   if (degreeprogrammes.length === 0) {
-    console.error("No programmeDegrees found!");
-    process.exit(1);
+    throw new Error("No degree programmes found!");
   }
   console.debug(`✅ ${degreeprogrammes.length} programmes found.`);
 
@@ -199,19 +129,24 @@ async function main() {
         ? programme.disciplines.discipline
         : [programme.disciplines.discipline];
       for (const d of arr) {
-        if (d.id && getTextDe(d.name))
-          if (d.area_of_study) {
+        if (d.id && getTextDe(d.name)) {
+          if (
+            d.area_of_study &&
+            d.area_of_study.id &&
+            getTextDe(d.area_of_study.name)
+          ) {
             area_of_study.push([
               d.area_of_study.id,
               getTextDe(d.area_of_study.name),
             ]);
           }
-        disciplines.push([
-          d.id,
-          getTextDe(d.name),
-          d.area_of_study?.id || null,
-        ]);
-        if (d.id) programme_disciplines.push([id, d.id]);
+          disciplines.push([
+            d.id,
+            getTextDe(d.name),
+            d.area_of_study?.id || null,
+          ]);
+          programme_disciplines.push([id, d.id]);
+        }
       }
     }
 
@@ -263,8 +198,6 @@ async function main() {
   }
 
   try {
-    await client.query("BEGIN");
-
     await batchInsert(client, "abschlussart", ["id", "name"], degrees, "(id)");
     await batchInsert(
       client,
@@ -367,16 +300,9 @@ async function main() {
       "(studiengang_id, standort_id)",
     );
 
-    await client.query("COMMIT");
     console.debug("Import succeeded!");
   } catch (err) {
-    await client.query("ROLLBACK");
-    console.error("Import failed, rollback executed.", err);
-  } finally {
-    await client.end();
+    console.error("Import failed.", err);
+    throw err;
   }
 }
-
-main().catch((err) => {
-  console.error("Unexpected error:", err);
-});
