+++ server/package.json
@@ -26,6 +26,7 @@
     "express-async-errors": "^3.1.1",
     "express-rate-limit": "^8.2.1",
     "pg": "^8.16.3",
+    "pg-copy-streams": "^7.0.0",
     "xml2js": "^0.6.2"
   },
   "devDependencies": {
@@ -35,6 +36,7 @@
     "@types/jest": "^30.0.0",
     "@types/node": "^24.10.1",
     "@types/pg": "^8.15.6",
+    "@types/pg-copy-streams": "^1.2.5",
     "@types/supertest": "^6.0.3",
     "@types/xml2js": "^0.4.14",
     "jest": "^30.2.0",
